requirements.txt
pyTelegramBotAPI>=4.14.0
flask>=3.0.0
requests>=2.31.0
pycryptodome>=3.19.0
PyJWT>=2.8.0
protobuf>=4.25.0
urllib3>=2.0.0
gunicorn>=21.2.0
